requirements.txt
PyMuPDF
pdfminer.six
PyPDF2
pdfplumber
PyYAML
pytest
pytest-cov
pytest-mock
camelot-py[cv]
ftfy
clean-text
spacy
rank_bm25
tiktoken
faiss-cpu
numpy
requests
streamlit
google-generativeai
openai
unidecode